Update README.md
It is an upgrade from my previous project using linear regression model. It is an implementation of a Convolutional Neural Network (CNN) for image classification, designed to classify images from two popular datasets: MNIST and CIFAR-10. CNNs are a type of deep learning model known for their exceptional performance in image-related tasks.